added allocation calculation

+++ backend/src/project/project.service.ts
@@ -2,20 +2,37 @@ import { toDTO } from '@app/runtime/util'
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common'
 import { InjectModel } from '@nestjs/mongoose'
 import { FilterQuery, Model } from 'mongoose'
-import { CreateProjectDto, ProjectDto, ProjectQueryDto } from './project.dto'
+import { CreateProjectDto, ProjectDto, ProjectEvent, ProjectQueryDto, ProjectStatus } from './project.dto'
 import { Project, ProjectDocument } from './project.schema'
+import { EventEmitter2 } from '@nestjs/event-emitter'
+import { RecordEventType } from 'src/task/task.dto'
 
 @Injectable()
 export class ProjectService {
-  constructor(@InjectModel(Project.name) private projectModel: Model<ProjectDocument>) {}
+  constructor(
+    private eventEmitter: EventEmitter2,
+    @InjectModel(Project.name) private projectModel: Model<ProjectDocument>,
+  ) {}
 
   toDto(doc: ProjectDocument): ProjectDto {
     return toDTO<ProjectDto>(doc)
   }
 
+  private emit(type: RecordEventType, project: ProjectDto | ProjectDocument) {
+    this.eventEmitter.emit('project.changed', {
+      projectId: project.projectId,
+      daoId: project.daoId,
+      type,
+    } as ProjectEvent)
+  }
+
   async create(projectDto: CreateProjectDto): Promise<ProjectDto> {
     const project = new this.projectModel(projectDto)
+
+    project.status = ProjectStatus.open
+
     await project.save()
+    this.emit('create', project)
     return this.toDto(project)
   }
 
@@ -27,7 +44,7 @@ export class ProjectService {
   async update(projectDto: ProjectDto): Promise<ProjectDto> {
     if (!projectDto) throw new BadRequestException()
     const { projectId } = projectDto
-    if (projectId) throw new BadRequestException()
+    if (!projectId) throw new BadRequestException()
     const project = await this.read(projectDto.projectId)
     if (!project) throw new NotFoundException()
 
@@ -38,12 +55,19 @@ export class ProjectService {
 
     const res = await this.projectModel.updateOne({ projectId }, updates)
     if (!res.matchedCount) throw new NotFoundException()
-
+    this.emit('update', project)
     return await this.read(projectId)
   }
 
   async delete(projectId: string): Promise<void> {
+    const project = await this.read(projectId)
+    if (!project) return
     await this.projectModel.deleteOne({ projectId }).exec()
+    this.emit('delete', project)
+  }
+
+  async deleteAll(): Promise<void> {
+    await this.projectModel.deleteMany({}).exec()
   }
 
   async find(query: ProjectQueryDto): Promise<ProjectDto[]> {
